example/pyproject.toml
[tool.poetry]
name = "example"
version = "0.1.0"
description = ""
authors = ["Automat z.s."]
license = "LGPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.7"
django-import-export = "*"
django = "*"
celery = "*"
django-author = "*"
redis = "*"
psycopg2-binary = "^2.8.4"

[tool.poetry.dev-dependencies]
django-admin-smoke-tests = "^0.3.0"
pudb = "^2019.2"
black = "^22.3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
